extract step into function
