+++ core/simulations/configs/vaultLiquidation.ts
@@ -1,33 +1,16 @@
-import {
-    warpTime,
-    resetNetworkAndSetupWallet,
-    setCollateralInVat,
-    addDaiToBalance,
-    addMkrToBalance,
-} from '../../helpers/hardhat';
+import { warpTime, resetNetworkAndSetupWallet, addDaiToBalance, addMkrToBalance } from '../../helpers/hardhat';
 import { Simulation } from '../types';
 import prompts from 'prompts';
-import COLLATERALS, { getCollateralConfigByType } from '../../src/constants/COLLATERALS';
+import COLLATERALS from '../../src/constants/COLLATERALS';
 import BigNumber from '../../src/bignumber';
-import {
-    changeVaultContents,
-    collectStabilityFees,
-    fetchVault,
-    openVault,
-    liquidateVault,
-    fetchVaultCollateralParameters,
-} from '../../src/vaults';
+import { collectStabilityFees, fetchVault, liquidateVault, fetchVaultCollateralParameters } from '../../src/vaults';
 import { HARDHAT_PUBLIC_KEY, TEST_NETWORK } from '../../helpers/constants';
-import getContract, {
-    getContractAddressByName,
-    getErc20Contract,
-    getJoinNameByCollateralType,
-} from '../../src/contracts';
-import { depositCollateralToVat, fetchCollateralInVat, withdrawCollateralFromVat } from '../../src/wallet';
-import { MAX, RAD_NUMBER_OF_DIGITS } from '../../src/constants/UNITS';
+import getContract from '../../src/contracts';
+import { RAD_NUMBER_OF_DIGITS } from '../../src/constants/UNITS';
 import { CollateralType } from '../../src/types';
 import { ethers } from 'ethers';
 import { randomBigNumber } from '../../helpers/hex';
+import createVaultForCollateral from '../steps/createVaultForCollateral';
 
 const getCollateralType = async () => {
     const { collateralType } = await prompts([
@@ -44,12 +27,6 @@ const getCollateralType = async () => {
         },
     ]);
     return collateralType;
-};
-
-const getLatestVault = async () => {
-    const cdpManager = await getContract(TEST_NETWORK, 'CDP_MANAGER', true);
-    const lastHex = await cdpManager.last(HARDHAT_PUBLIC_KEY);
-    return new BigNumber(lastHex._hex).toNumber();
 };
 
 const getVaultLimits = async (collateralType: CollateralType) => {
@@ -100,71 +77,10 @@ const simulation: Simulation = {
         {
             title: 'Create the auction',
             entry: async context => {
-                console.info('Setting collateral in VAT...');
-                const collateralType = context.collateralType;
-                await setCollateralInVat(collateralType, context.collateralOwned);
-                let balance = await fetchCollateralInVat(
-                    TEST_NETWORK,
-                    HARDHAT_PUBLIC_KEY,
-                    collateralType,
+                const latestVaultId = await createVaultForCollateral(
+                    context.collateralType,
+                    context.collateralOwned,
                     context.decimals
-                );
-                if (!balance.eq(context.collateralOwned)) {
-                    throw new Error(
-                        `Unexpected vat balance. Expected: ${context.collateralOwned.toFixed()}, Actual: ${balance.toFixed()}`
-                    );
-                }
-                console.info(
-                    `Vat Collateral ${collateralType} balance of ${HARDHAT_PUBLIC_KEY} is ${context.collateralOwned.toFixed()}`
-                );
-                console.info('Opening the vault');
-                await openVault(TEST_NETWORK, HARDHAT_PUBLIC_KEY, context.collateralType);
-                console.info('Extracting collateral');
-                const addressJoin = await getContractAddressByName(
-                    TEST_NETWORK,
-                    getJoinNameByCollateralType(context.collateralType)
-                );
-                const collateralConfig = getCollateralConfigByType(context.collateralType);
-                const tokenContractAddress = await getContractAddressByName(TEST_NETWORK, collateralConfig.symbol);
-                const contract = await getErc20Contract(TEST_NETWORK, tokenContractAddress, true);
-                await contract.approve(addressJoin, MAX.toFixed(0));
-                console.info('Max allowance given out');
-                await withdrawCollateralFromVat(TEST_NETWORK, HARDHAT_PUBLIC_KEY, context.collateralType, undefined);
-                const token = await getErc20Contract(TEST_NETWORK, tokenContractAddress);
-                const balanceHex = await token.balanceOf(HARDHAT_PUBLIC_KEY);
-                balance = new BigNumber(balanceHex._hex).shiftedBy(-context.decimals);
-                if (!balance.eq(context.collateralOwned)) {
-                    throw new Error(
-                        `Unexpected wallet balance. Expected ${context.collateralOwned.toFixed()}, Actual ${balance.toFixed()}`
-                    );
-                }
-                console.info(
-                    `Wallet ${HARDHAT_PUBLIC_KEY} has ${context.collateralOwned.toFixed()} of token ${tokenContractAddress}`
-                );
-                console.info('Depositing Collateral to vault');
-                const latestVaultId = await getLatestVault();
-                let vault = await fetchVault(TEST_NETWORK, latestVaultId);
-                await depositCollateralToVat(
-                    TEST_NETWORK,
-                    vault.address,
-                    vault.collateralType,
-                    context.collateralOwned
-                );
-                console.info('Adding collateral to Vault');
-                vault = await fetchVault(TEST_NETWORK, latestVaultId);
-                const drawnDebtExact = context.collateralOwned
-                    .multipliedBy(vault.minUnitPrice)
-                    .dividedBy(vault.stabilityFeeRate);
-                const drawnDebt = new BigNumber(
-                    drawnDebtExact.toPrecision(drawnDebtExact.e || 0 + 1, BigNumber.ROUND_DOWN)
-                );
-                console.info(drawnDebt.toFixed(), vault.minUnitPrice.toFixed(), drawnDebtExact.toFixed());
-                await changeVaultContents(TEST_NETWORK, latestVaultId, drawnDebt, context.collateralOwned);
-                const vaultWithContents = await fetchVault(TEST_NETWORK, latestVaultId);
-                console.info(
-                    `Vault's contents: ${vaultWithContents.collateralAmount.toFixed()} of collateral, ${
-                        vaultWithContents.initialDebtDai
-                    } of debt`
                 );
                 return { ...context, latestVaultId };
             },
